foreign key bobot

+++ database/migrations/8_create_sub_cpmk_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
     public function up()
     {
         Schema::create('sub_cpmk', function (Blueprint $table) {
-            $table->integer('sub_cpmk')->primary;
+            $table->bigIncrements('sub_cpmk')->primary;
             $table->string('Deskripsi_Sub_CPMK');
         });
     }
